requirements.txt
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2
django-cors-headers==2.5.2
django-rest-framework==0.1.0
djangorestframework==3.9.2
idna==2.8
pytz==2019.1
requests==2.21.0
sqlparse==0.3.0
Unidecode==1.0.23
urllib3==1.24.2
whitenoise==4.1.2
